create FilmDetailsResolver
